fix: crashes due to missing fields in simctl output (#1538)
Fixes #1517

For older runtimes, the `deviceTypeIdentifier` field is missing in the
output of `simctl`, causing crashes when trying to read the devices in
the system-wide device set.
Previously this wasn't an issue, because we only accessed the devices
created by us, which always included that information.
However, in #1471 we read the system-wide devices to choose a
destination to pass to `xcodebuild`, resulting in build errors when
trying to access the non-existent property.
This PR fixes that.