feat: Device Settings test coverage (#1526)
This PR introduces **automated tests for Device Settings**, and extends
the existing suite of **Radon tools tests**.

Additionally, the test related to changing the device appearance does
not pass,
but this is expected since the functionality is not working in this
commit.

---

### How Has This Been Tested

Tests have been executed:  
- On **GitHub Actions** (iOS)  
- Locally on **Android** and **iOS**  

To run tests on GitHub CI:  
Start Workflow on branch `@KeyJayY/DeviceSettingsTests` from the Actions
tab in GitHub.

To run tests locally:  
`npm run prepare-and-run-tests -- react-native-80`  
(or any other supported test app).  

To run only specific tests:  
`npm run prepare-and-run-tests -- react-native-81 6 7 8`  
executes tests 6, 7 and 8 (tests that were modified). Any combination is
supported.

In `utils/configuration.js`, you can switch between Android and iOS test
runs.

Local VM can also be used, to run tests on VM you can use command:
`npm run run-tests-on-VM -- react-native-81 6 7 8`

> Note: Tests related to changing the device appearance (light/dark
mode) do not pass on iOS on this commit. This is expected, as the
functionality is not working correctly in this version.

---

### How Has This Change Been Documented

Not documented yet.