Use latest version of source-map instead of beta published 7 years ago (#1568)
For an unknown reason while updating the source-map module version, we
installed version 0.8.0-beta.

This version turns out not to be actively maintained though and doesn't
look like a successor of 0.7. The 0.8.0-beta.0 was published 7 years ago
while 0.7 is actively maintained with fixes and improvements every
month.

As we are experiencing some crashes in the source map code, I figured
it'd be better that we are using a supported version instead of an
abandoned beta.

Source maps are currently only used in our custom debugger
implementation.

### How Has This Been Tested: 
1. Run the app with custom config and set `"useCustomJSDebugger": true`
– this way the
2. Place some breakpoints to see them work, and check debug console for
decpihered line numbers