feat: Expo Router test coverage (#1562)
This PR introduces a new test suite for Expo router testing. It also
improves the Radon Preview test. Additionally, the [Test
apps](https://github.com/software-mansion-labs/radon-ide-test-apps) were
updated to be compatible with the new tests.

### How Has This Been Tested

- Tested locally (iOS and Android)
- Tested on GitHub Actions (iOS)

To reproduce, run one of the following commands:  
`npm run run-tests-on-VM -- expo-54 7 10`  
`npm run prepare-and-run-tests -- expo-54 7 10`  

### How Has This Change Been Documented

If applicable (for UI changes or new features), provide a link to the PR
that updates the relevant documentation pages.