refactor: device settings (#1567)
This is almost an end in our quest to get rid off updating state using
the old `eventLinstener` mechanic.
This time we remove `deviceSettingsChanged` event and add deviceSettings
to the new state management solution.
At the same time we choose to locate them in `workspaceConfiguration` so
the user might share the settings with team members or other devices.

The move to `workspaceConfiguration` would make them bloated so we
refactor them to have "subsections".

### How Has This Been Tested: 

- run test applications and test if sttings are applied, and persisted. 

### How Has This Change Been Documented:

internal (?) there is a change to the settings structure but it is self
documented