feat: Schedule automatic runs and code refactor  (#1571)
This PR adds a scheduled job to run tests automatically at 3:27 AM. It
also includes a code refactor, with the main change being the separation
of logic previously in `radonInteractionsServices.js` into separate
files. This change increases readability.

### How Has This Been Tested: 

Tested locally on iOS and Android

to reproduce you can use one of the following commands:
`npm run prepare-and-run-tests`
`npm run run-tests-on-VM`


### How Has This Change Been Documented:

If applicable (for UI changes or new features) provide a link to a PR
that updates the relevant documentation pages.