fix: debugger overlay not showing while status not running (#1573)
This solves an issue that could happen if the user forces the preview
screen while debugger was enabled, not allowing to continue the debugger
in any way.

This PR also resets the `showPreviewRequested` flag on every status
change which makes it go away on reloads.

### How Has This Been Tested: 

- run radon on test commit (c278d25f3a977f5aff5fb4d21e384d1488ea9406)
with information about the device session status obscured and check if
debugger works correctly

### How Has This Change Been Documented:

internal