feat: report Metro initialization errors to the user (#1578)
Currently, errors when starting the metro server are not handled and
reported to the user correctly.
It wasn't a visible issue before, since we didn't expect the metro
server to exit early. However, after adding the `metroPort` option,
metro failing due to listening on a taken port is an expected error, and
should be reported to the user.
This PR introduces a new "MetroError" type that can be reported to the
frontend and adds an error alert that is shown for that error.

### How Has This Been Tested: 
- test on bare `metro` and `expo`
- set the `metroPort` option to some taken port
- run the app
- an error alert should appear
- use auto reload or any reload option except "Reload JS"
- the error alert should appear