feat: support prebuilding expo dev client applications (#1576)
This PR adds automatic detection of expo dev client setups. 

additionally we use this opportunity to clean up the `/lib` section of
our repository as it was a bit overblown for some time and we add a new
file to it.

additionally additionally we fix the way we import expo cli in a path
that was missed in
https://github.com/software-mansion/radon-ide/pull/1522

### How Has This Been Tested: 

- create a new expo 54 test app and 

### How Has This Change Been Documented:

internall-ish?