Update branch of vscode-js-debug after vscode-js-debug/pull/8 got merged (#1583)
In #1580 we bumped version of vscode-js-debug, but since the changes
weren't merged at that point, the submodule was pointing to branch
corresponding to
https://github.com/software-mansion-labs/vscode-js-debug/pull/8

I should've updated the submodule to point to the main branch instead
after merging but forgot to do so. This PR fixes this and makes the
vscode-js-debug submodule point to main branch on the fork repo.