Fix crash on IDE.dispose (#1588)
After #1386 the IDE.dispose method started throwing an error that
c.dispose is not a function.

The issue was in the "mock" channel introduced in that PR. The code used
`as` operator to cast incomplete type into channel, but it was missing a
lot of methods (including the dispose method) that some parts of the
codebase could use.

This PR adds a complete LogChannel prototype for the mock
implementation.

### How Has This Been Tested: 
1) Open IDE panel
2) Use the "switch off" button to turn off the panel when in sidebar
mode or just close it the tab in the "tab mode"
3) Before this change we'd get an error that c.dispose is not a
function, now this error is gone