Fix back and home behavior in UrlBar (#1595)
This PR changes the logic behind integration with URL bar, back and home
buttons in order to address multiple issues that we have:
1) Currently, the back button is completely disabled in non expo-router
projects, it still occupies the UI.
2) Home button reloads the app, which just isn't intuitive. On top of
that it also leaves the history polluted which also isn't what people
expect from the browser.
3) Back button doesn't work when preview is opened (doesn't navigate
back)
4) The preview URL is displayed in the URL bar after "home" button gets
pressed and despite the app showing the main screen not the preview.

### Top-level approach for addressing these points:
1) We should use the back button to allow for going back from preview
mode.
2) Home button should just wind back the navigation stack, and on
non-router apps it should reset the app to initial screen instead of
reloading JS which sometimes takes a while. Finally, home button should
reset navigation history and give you a fresh start.
3) Using back on preview should go back in the history (to previous
preview or to the app in its last state ideally)

### Fundamental issues in current approach
While investigating the url handling logic I found out about a couple of
fundamental issues that this PR fixes first:
1) when opening previews, because expo router is registered as
navigation plugin, the preview container would try to use expo router
plugin for handling navigation events.
2) using navigate("/") in expo router would in certain setting perform a
push transition which just look very weird when you want to go "HOME".

### Notable changes in this PR
Below I document the key changes made in this PR:
1) Changing the logic in `sendNavigationChange:` before it'd only
trigger a change if there was already some previous route set. This led
to the initial routes not being reported. Because of other issues that
eventually triggered following updates, we missed this problem before.
2) expo router plugins now have special handling for __HOME__ event. In
v5 and v6 we use new `dismissTo` method as it optimistically allows us
to go back to "/" with a single command. On older versions we use
`dismissAll` which always require a follow up `navigate("/")` call.
3) We change the `initialProps` keys to always use radon specific
prefixes (i.e. `__radon_previewKey`) – this is needed as we read and
propagate initial props down to app components. This way we avoid name
collision with apps
4) In wrapper, we capture the initial `appKey` (scene name). This is
needed such that we can actually close the preview. Before we were using
"main" which is the key used in expo-router apps, but we needed this to
work on non-expo-router apps too.
5) We delete `__RNIDE_onMount` initial prop handling as it wasn't being
used anywhere.
6) The goHome handling in project now always sends "openNavigation"
request. We use special key `__HOME__` similarily to how we use
`__BACK__`. Home handling is moved from extension logic into the wrapper
and plugins. In addition, we clear the history when "home" is requested.
7) We extend `NavigationItem` type to include `canGoBack` flag. In
expo-router plugins, the flag is propagated based on
`router.canGoBack()` call, we also set it to `true` for previews. Later,
this flag is used to determine whether back arrow next to the Url bar is
active or not. We further extend the type to also accept undefined as
display name, in that case the frontend will show the default home route
label (it is "/" in our case now but it is spread all over the codebase
right now and didn't want to change that)
8) We add special-handling for the case when HOME is requested from
non-expo-router app. In that scenario we force re-render the application
using a dummy component that is rendered in between. Apparently
AppRegistry doesn't have a method to force re-mounting.

### Unsolved problems in this area
Some problems that are still unsolved but out of scope for this change:
1) when opening two previews one after another, back button would still
just go to the main app and not the the last preview.
2) using modal routes from the history would often push a new one
instead of navigating back to one that is already open on the stack.
3) if we open preview after navigating first to some non-root route,
back button from preview should get us back to that route while it takes
us to the root route now.
4) When you use "home" on react-conf-app it results in the url bar
dropdown presenting multiple "/" routes. This is a bug in our route
resolution code which for some reason returns multiple "/" routes but
doesn't show some legal routes like "/info". This problem was mitigated
before, because there was always a "/" route in the history which was
causing these invalid routes to be filtered out before presenting.

### How Has This Been Tested: 
1) I used react-conf-app to test expo router integrations, in
particular: back button after opening preview, home button after opening
preview, home button after navigating deep into router hierarchy, back
button at different stages
2) I used react-native-81 to test non-expo router setup in particular:
going back from preview, home button from preview, home button after
changing the screen