feat: skip waiting for fingerprint on autoreload (#1600)
Currently, when "Auto Reload" is requested, we calculate the project's
fingerprint to check, if we need to rebuild the app. This can make the
reloads costly on larger projects.
This PR changes the behavior to instead check the "isUsingStaleBuild"
state, which is set on fingerprint mismatch, which allows us to skip the
calculation and start reloading immediately.
We also immediately check if "isUsingStaleBuild" should be set, so that
if the user auto-reloaded before some native project changes were picked
up by Radon, they will be prompted to reload again as soon as the
fingerprint calculation completes.

### How Has This Been Tested: 
- open an app which takes a long time to calculate the fingerprint, for
example by setting the `fingerprintCommand` launch configuration to
```
sleep 10; echo 'A';
```
- trigger autoreload
- the app should restart quickly, without waiting for the fingerprint to
calculate