feat: support pnpm workspaces (#1594)
This PR changes how we resolve/require/override modules in application
dependencies.

Previously we assumed that if something is a dependency of
`react-native` it can be found in the applications `node_modules` as it
is a how `yarn` and `npm` do it, unfortunately while using some monorepo
setups (`pnpm workspace` in priticular) only the dependencied of the
application are linked in its node modules. ( You can read more about it
[here](https://pnpm.io/faq#pnpm-does-not-work-with-your-project-here. )
)

To solve this problem we require the dependencies from the dependency
directory instead of the app directory.

additionally after introducing this concept to "react-native"
dependencies it felt silly to still do it by hand for "expo", as
introduced in #1522, so we move those imports/overides to the new
solution as well.

relates to: #1579

### How Has This Been Tested: 

- Clone the
[expo-monorepo-example](https://github.com/byCedric/expo-monorepo-example)
repo.
- run it in radon

### How Has This Change Been Documented:

work in progress