Avoid concurrent initialization of devtools connection over CDP binding (#1617)
This problem fixes an issue with concurrent initialization of the
CDP-backed devtools connection.

Before this change, we would call the connection initialization code
from both `onJSDebugSessionStarted` and `onScriptParsed` when the main
bundle got loaded. We had a check for whether the connection was already
set to avoid initializing it twice. However, since the initialization
was async, it could happen that both of these calls would advance past
the check. This eventually resulted in us emitting two connect events
and shutting down a functioning connection.

While in practice we'd expect that re-creating the agent should work
alright and we should just get a second instance of an agent. It appears
however that interleaving the binding call and evaluate call may confuse
the devtools backend code too much and results in the agent connection
being dropped entirely. When testing this, if the second initialization
call happens after some delay, this work correctly.

However, we still want to avoid two connections being initialized and
therefore we need to prevent the concurrent initialization anyway.

The consequence of the above bug was that the devtools connection would
shut down shortly after being initialize and as a result we wouldn't be
able to receive communication over the devtools channel (i.e. the
network inspector calls).

So one way we could reproduce the problem was on a simple expo 52 dev
client app. After starting the app and opening the network panel, we
woudn't be able to see network requests. However, after "reload JS" the
request would start appearing as the connection was re-established.

### How Has This Been Tested: 
1. I could only reproduce the issue on small expo-52 dev-client setups
2. `npx create-expo-app@latest --template default@52`
3. `npx expo install expo-dev-client`
4. Add button that performs network fetch
5. before the change after launching the app and opening network
inspector you couldn't see requests coming until you click fast refresh
6. after this change the requests are showing up after clean start