fix: CPU Profiles aren't saved correctly (#1625)
This PR fixes CPU profiles not saving after #1580 
- I (and other team members) could not get the CPU Profiler behaviour
described in #1580 to happen, and `vscode-js-debug` doesn't save the
profiles as described.
By reading the code, it turned out the profile file is save to `tmpdir`
and the path to that file is sent in the DAP custom event fired when
profiling starts. We can use that path to restore the previous behaviour
of asking the user where the file should be stored and automatically
opening it
- After that, `vscode-js-debug` could not annotate the sources for the
profile correctly.
The issue is caused by the `Profiler.Stop` result always referring to
`scriptId: '0'`, while the debugger never emitted a `scriptParsed` event
for that `scriptId`. This was resolved for now by replacing the
`scriptId` field with the last loaded main bundle's id in the
`ProxyDelegate`.

### How Has This Been Tested: 
- run an application
- start and stop the CPU profiler
- the file picker dialog should appear
- after saving the profile, the JS function entries in the profile
should link to files in the filesystem instead of the Metro server (the
profile also includes functions implemented within Hermes's native code
as well as Host Functions, which aren't mapped by the source maps and
still show `localhost` urls, but fixing that is outside the scope of
this PR)