Don't enable Radon AI on other VSCode forks (#1624)
Radon AI causes some additional friction for the users on Cursor (and
other VSCode forks) as it requires that we update mcp.json file every
time the editor starts. This doesn't happen on VSCode as since version
1.100 there's an API that allows our extension to register the MCP
server dynamically without updating the configuration file.

As a consequence, Radon AI may get in a way of users who doesn't want to
use it.

On top of the mcp json file being generated in the user's workspace,
Cursor recently started asking whether to enable MCP servers added
there. Because the mcp.json record has a port number that changes with
every start, Radon AI generates prompts with every start as there's a
new, unique record added to mcp.json each time.

Below are highlights of what this PR is changing:
1. We're changing the setting name to RadonAi.enabled and use 'enum'
based setting instead of boolean. Boolean setting doesn't allow us to
tell when the setting is not set (default value). It is important as we
use different defaults for VSCode and Cursor now. Users who had turned
the setting off on VSCode will need to turn it off again, but since
Radon Ai isn't intrusive on VSCode as it is on Cursor I don't think this
should be an issue.
2. We restructure the code around initialization: we move functions and
types for detecting VSCode vs Cursor to general utility file (was under
"ai" before). We refactor RadonAi loading scheme such that it reads from
the config and decides about the defaults afterwards.
3. We delege all the code responsible for defining radon ai settings in
our configuration scheme. This wasn't used anywhere and it was just a
lot of boilerplate
4. We update the registration logic such that it can respond to changes
in the setting in runtime. This allows us to modify the setting
description to no longer include the note about resetting.

### How Has This Been Tested: 
1. Run extension on Cursor
5. Don't get "enable MCP server" prompt
6. Change config to enable Radon AI
7. Get the prompt there and use Radon in the chat.
8. Do the same in cursor – verify Radon AI is available w/o changes to
the setting