fix: radon connect  (#1635)
This PR fixes a regression introduced in #1550 which removed an option
to pass `watchFolders` to the metro class directly. Normally we really
on metro process handing as this information during the set up process,
but when we use radon connect we do not have an option of capturing
metro config like in a normal case, so we need to add the `watchFolders`
information by hand.

### How Has This Been Tested: 

- run expo 54 test app from terminal 
- use radon connect 

### How Has This Change Been Documented:

internal