feat: make `usePrebuild` option more important then automatic heuristic  (#1638)
In our recent changes to the prebuild system, we added an automatic
detection to check if the project uses `dev-client` and force a prebuild
if it does. This however made us ignore the old `usePrebuild` option in
such a scenarion not allowing a user to opt in into a manual prebuild by
default.

### How Has This Been Tested: 

- run `social-app` 
- set the `usePrebuild` option to false
- after it works reload the whole of radon ide

### How Has This Change Been Documented:

not different from the 1.11.2