Update custom build path parser to fix issues with dot directory parsing (#1644)
This PR fixes the logic behind the custom build file path parsing.

The previous regex that we used was falling short when there was a dot
(".") in the path followed by an empty space, for example:
`/Users/krzys/.builds/My App.app`

This PR updates the parsing regex to be more resilient and also consider
the cases where quotes are put around the paths (which is often the case
when paths include spaces).

On top of that the package contributes a number of unit tests for the
new parser.

### How Has This Been Tested: 
1. Run tests