Add .watchmanconfig file to lib folder (#1650)
Radon adds a custom `watchFolder` to metro configuration via metro
config overwrite mechanism. This results in watchman picking up that
folder as one of its roots to watch. However, some watchman variations
doesn't like the fact that they are unable to locate the config instead
of falling back to a default config. This PR adds a watchman config in
the `lib/` folder, which is the one that we point `watchFolders` to
resulting in watchman process picking up the right root to watch.

### How Has This Been Tested: 
1. In development: run any app, use `watchman watch-list` in terminal –
see it shows `lib/` folder on the watch list instead of the root folder
where this project is checked out. (to clean up the list of watched
folders you can run `watchman watch-del-all`)
2. Package extension to verify that the .watchmanconfig file is in the
distribution package: run `npm run vscode:package` and then unzip the
produced vsix file.