Move blacklist overwrites in metro.config (#1649)
In #1581 we added metro config blacklist overwrites in order to prevent
fast refresh from getting triggered when typedefinition files are
modified. However, the overwrites were only added to one conditional
branch that was testing if the config.resolver was set. We want to apply
the overwrites regardless of that and hence this PR moves that code
outside of the condition.

### How Has This Been Tested: 
1) open bsky app, make sure it loads
2) just wait for a bit. Before, every 15 seconds or so, you'd see the
"Fast refresh completed" notification which was due to the fingerprint
being run. After this change, this no longer happens. App loads normally
and responds to refresh calls from normal JS/TS files.