fix: metro config in non flat dependency structures (#1655)
This PR updates the logic of how we adapt metro config for expo projects
to support `pnpm workspace`.
Since Expo 54 the `metro-config` module is no longer imported as a
`react-native` dependency but is located in `@expo/metro/metro-config`
instead. That make our old method of adapting it pass with no effect if
`node_modules` do not have a flat structure.
 
### How Has This Been Tested: 

- run test [pnpm
monorepo](https://github.com/byCedric/expo-monorepo-example) to verify
it worked
- run `expo-54`, `expo-53` and `expo-52-prebuildwithplugins` test apps 

### How Has This Change Been Documented:

internal