Use initialUrl option to launch expo go and devclient projects on iOS (#1662)
A while ago (in SDK 52) expo added an option that allows for Expo Go and
Dev Client to pass the link as the process arguments rather than using
open URL: https://github.com/expo/expo/pull/33182

This has two main benefits:
1) we no longer need to fix the permission setting for the URL to open
automatically without a dialog
2) we can monitor the app process output (when launching via deeplink we
don't get the stdout from the app process)

This PR adapts this mechanism by first detecting the SDK version and
using it conditionally.

1) we pass appRoot to launchApp because it is needed in order to detect
the version. While this new parameter doesn't fit super well it seems to
be the option.
2) we use appRoot to detect expo version using the package discovery
(similar to how we check for min version)
3) we append `--initialUrl` with the deeplink as launch arguments for
Expo SDK 52+
4) we're adding additional logs in that place to make it easier to
detect issues in the launch process

### How Has This Been Tested: 
1) run a couple of different configurations of projects: expo devclient
on SDK 52, 53 and 51, expo go on 53 and 51.
2) Verify the project launches and that it prints that the deeplink is
used on SDK 51.
3) On SDK 53 you should see "Radon IDE (Simulator Logs)" among output
channels