feat: network inspector - ignore internal React Native messages (#1651)
### Description

This PR changes the network inspector behaviour, so that the the
internal React Native network messages are ignored and not shown to the
user. The filtering is done on the extension-site in the
`BaseNetworkInspector` abstract class, at the point where the message is
broadcast to the webviews.

### How Has This Been Tested: 

Features were tested in local development environment of extension as
follows:
- open radon-ide/packages/vscode-extension
- [ON FIRST RUN] build simulator-server locally as follows:
- in `/packages` run: `git submodule update --init --recursive
packages/simulator-server`
  - step into simulator-server folder: `cd simulator-server`
- install dependencies as given in README.md and build
simulator-server-macos file locally using: `cargo build --release `
  - step into extension folder: `cd /packages/vscode-extension`
- run: `npm run build:sim-server-debug` or copy built binary from
simulator-server into `dist` folder
- run the extension locally using Run and Debug menu -> Run Extension
- open an application supporting Radon IDE in editor - example of app
supporting rotation: radon-ide-test-apps/react-native-80
- run Radon IDE extension panel
- *test whether the internal messages (such as the ones sent by opening
the `LogBox`, for example caused by unexpected error in the React Native
app) are being properly ignored in both implementations of network
inspectors*
### How Has This Change Been Documented:

Not applicable.