fix: network inspector - optimise syntax highlighting on the frontend (#1652)
### Description

This PR optimises syntax highlighting on the frontend in order to reduce
freezes when using network inspector

Optimisations:
- introduced `isActive` predicate to `HighlightedCodeBlock`, to perform
syntax highlighting only when the proper tab is active, which was not
the case so far - even if the user was looking at the `Headers` tab, the
code highlighting was still done on the main thread
- introduced `HighlighterContext`, whose main purpose is to keep cache
of highlighted code at the top level of the application. Based on hash
of code (dependent on requestId, code itself, theme and language) we
keep the HTML of the highlighted code in memory.
- reduced the `MAX_HIGHLIGHT_LENGTH` from `100_000` to `65_000` based on
tests. To be discussed.
- removed the `react-shiki` package in favour of
[`shiki`](https://github.com/shikijs/shiki), because caching with
`react-shiki` was not possible
- 
### How Has This Been Tested: 

Features were tested in local development environment of extension as
follows:
- open radon-ide/packages/vscode-extension
- [ON FIRST RUN] build simulator-server locally as follows:
- in `/packages` run: `git submodule update --init --recursive
packages/simulator-server`
  - step into simulator-server folder: `cd simulator-server`
- install dependencies as given in README.md and build
simulator-server-macos file locally using: `cargo build --release `
  - step into extension folder: `cd /packages/vscode-extension`
- run: `npm run build:sim-server-debug` or copy built binary from
simulator-server into `dist` folder
- run the extension locally using Run and Debug menu -> Run Extension
- open an application supporting Radon IDE in editor - example of app
supporting rotation: radon-ide-test-apps/react-native-80
- run Radon IDE extension panel and open network inspector
- *fetch different large files, which can be highlighted, see if caching
and colouring works as expected without causing freezes*
### How Has This Change Been Documented:

Not applicable.